+++ CHANGELOG.md
@@ -2,6 +2,8 @@
   - Warn on duplicate fields (see #283)
   - Always render `cabal-version` as `x.y` instead of `>= x.y` so that `cabal
     check` does not complain (see #322)
+  - Extend `build-tools` so that it subsumes Cabal's `build-tools` and
+    `build-tool-depends` (see #254)
 
 ## Changes in 0.29.7
   - Expose more stuff from `Hpack.Yaml` so that it can be used by third parties